backend/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
openai==0.27.2
python-dotenv==1.0.0
